backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
pillow==10.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
google-cloud-firestore==2.14.0
psycopg2-binary
pyotp
qrcode
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
